Conflicts:
	src/Core/Builder/ConstantGrammar.hpp
	src/Core/Builder/SymbolTable.hpp
	src/Core/Builder/VariableGrammar.hpp
	src/Core/Filter/ConstSymbolVisitor.cpp
	src/Core/Filter/ConstSymbolVisitor.hpp
	src/Core/Filter/SymbolVisitor.cpp
	src/Core/Filter/SymbolVisitor.hpp
	src/Core/IR/Symbol.cpp
	src/Core/Utility/Printer.cpp

+++ src/Core/IR/PlaceholderType.cpp
@@ -1,14 +1,21 @@
-#include <mirv/Core/Builder/MakeExpression.hpp>
-#include <mirv/Core/IR/PlaceholderType.hpp>
 // FIXME: Why is this necessary?
 #include <mirv/Core/IR/Module.hpp>
+#include <mirv/Core/IR/Function.hpp>
+#include <mirv/Core/IR/Variable.hpp>
+#include <mirv/Core/IR/GlobalVariable.hpp>
+#include <mirv/Core/Builder/MakeExpression.hpp>
+#include <mirv/Core/IR/PlaceholderType.hpp>
+#include <mirv/Core/Builder/ExpressionGrammar.hpp>
+
+#include <cstdint>
 
 namespace mirv {
   Placeholder::Interface::BitSizeType
   Placeholder::Interface::bitsize(void) const
   {
     BitSizeType size =
-      Builder::makeExpression(0ull, this->parent<Symbol<Module> >());
+      Builder::makeExpression(static_cast<std::uint64_t>(0ull),
+                              this->parent<Symbol<Module> >());
     return size;
   }
 }
